+++ components/com_crowdfunding/views/discover/view.html.php
@@ -45,6 +45,8 @@ class CrowdfundingViewDiscover extends JViewLegacy
 
     protected $pageclass_sfx;
 
+    protected $helperBus;
+
     public function display($tpl = null)
     {
         $this->option     = JFactory::getApplication()->input->getCmd('option');
@@ -53,12 +55,11 @@ class CrowdfundingViewDiscover extends JViewLegacy
         $this->items      = $this->get('Items');
         $this->pagination = $this->get('Pagination');
 
-        // Get params
-        $this->params = $this->state->get('params');
-        /** @var  $this->params Joomla\Registry\Registry */
+        $this->params      = $this->state->get('params');
 
-        $this->numberInRow       = (int)$this->params->get('items_row', 3);
-        $this->items             = CrowdfundingHelper::prepareItems($this->items, $this->numberInRow);
+        $this->numberInRow = (int)$this->params->get('items_row', 3);
+
+        $this->prepareItems($this->items);
 
         // Get the folder with images
         $this->imageFolder = $this->params->get('images_directory', 'images/crowdfunding');
@@ -68,7 +69,7 @@ class CrowdfundingViewDiscover extends JViewLegacy
         $this->amount = new Crowdfunding\Amount($this->params);
         $this->amount->setCurrency($currency);
 
-        $this->displayCreator = (bool)$this->params->get('integration_display_creator', true);
+        $this->displayCreator = (bool)$this->params->get('display_creator', true);
 
         // Prepare social integration.
         if ($this->displayCreator !== false) {
@@ -161,5 +162,20 @@ class CrowdfundingViewDiscover extends JViewLegacy
         }
 
         $this->document->setTitle($title);
+    }
+
+    private function prepareItems($items)
+    {
+        $options   = array();
+
+        $helperBus = new Prism\Helper\HelperBus($items);
+        $helperBus->addCommand(new Crowdfunding\Helper\PrepareItems());
+
+        // Count the number of funders.
+        if (strcmp('items_grid_two', $this->params->get('grid_layout')) === 0) {
+            $helperBus->addCommand(new Crowdfunding\Helper\PrepareItemFunders(JFactory::getDbo()));
+        }
+
+        $helperBus->handle($options);
     }
 }